# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Loan Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProduct"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /loans/{id}/emi:
    post:
      operationId: calculateEmi
      tags: [loans]
      summary: Calculate EMI for a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmiCalculationRequest"
      responses:
        "200":
          description: EMI calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmiCalculationResult"
  /applications:
    get:
      operationId: listApplications
      tags: [loans]
      summary: List user's loan applications
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoanApplication"
    post:
      operationId: createApplication
      tags: [loans]
      summary: Create a loan application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [loans]
      summary: Get loan application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [loans]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
...
  /admin/loans:
    post:
      operationId: createLoanProduct
      tags: [admin]
      summary: Create a new loan product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanProductRequest"
      responses:
        "201":
          description: Loan product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProduct"
  /admin/loans/{id}:
    put:
      operationId: updateLoanProduct
      tags: [admin]
      summary: Update a loan product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanProductRequest"
      responses:
        "200":
          description: Updated loan product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProduct"
    delete:
      operationId: deleteLoanProduct
      tags: [admin]
      summary: Delete a loan product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required:
        - message
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
        - message
        - success
...
    LoanProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [home, personal, business, education, vehicle, gold]
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
        minTenureMonths:
          type: integer
        maxTenureMonths:
          type: integer
        baseInterestRate:
          type: number
        processingFeePercent:
          type: number
        eligibilityCriteria:
          type: array
          items:
            type: string
        documentsRequired:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - description
        - minAmount
        - maxAmount
        - minTenureMonths
        - maxTenureMonths
        - baseInterestRate
        - processingFeePercent
        - eligibilityCriteria
        - documentsRequired
        - isActive
        - createdAt
    EmiCalculationRequest:
      type: object
      properties:
        loanAmount:
          type: number
        tenureMonths:
          type: integer
        interestRate:
          type: number
      required:
        - loanAmount
        - tenureMonths
        - interestRate
    EmiCalculationResult:
      type: object
      properties:
        monthlyEmi:
          type: number
        totalAmount:
          type: number
        totalInterest:
          type: number
        processingFee:
          type: number
        loanAmount:
          type: number
        tenureMonths:
          type: integer
        interestRate:
          type: number
      required:
        - monthlyEmi
        - totalAmount
        - totalInterest
        - processingFee
        - loanAmount
        - tenureMonths
        - interestRate
    LoanApplication:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        loanProductId:
          type: integer
        loanProductName:
          type: string
        loanAmount:
          type: number
        tenureMonths:
          type: integer
        monthlyEmi:
          type: number
        status:
          type: string
          enum: [pending, kyc_review, document_review, underwriting, sanctioned, rejected, locked, disbursed]
        agentStage:
          type: string
          enum: [sales, verification, underwriting, sanction, locking]
          nullable: true
        lockingId:
          type: string
          nullable: true
        creditScore:
          type: integer
          nullable: true
        remarks:
          type: string
          nullable: true
        applicantName:
          type: string
        applicantEmail:
          type: string
        applicantPhone:
          type: string
        annualIncome:
          type: number
          nullable: true
        employmentType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - loanProductId
        - loanProductName
        - loanAmount
        - tenureMonths
        - monthlyEmi
        - status
        - applicantName
        - applicantEmail
        - applicantPhone
        - createdAt
        - updatedAt
    CreateApplicationRequest:
      type: object
      properties:
        loanProductId:
          type: integer
        loanAmount:
          type: number
        tenureMonths:
          type: integer
        monthlyEmi:
          type: number
        annualIncome:
          type: number
        employmentType:
          type: string
      required:
        - loanProductId
        - loanAmount
        - tenureMonths
        - monthlyEmi
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, kyc_review, document_review, underwriting, sanctioned, rejected, locked, disbursed]
        agentStage:
          type: string
          enum: [sales, verification, underwriting, sanction, locking]
          nullable: true
        remarks:
          type: string
          nullable: true
        lockingId:
          type: string
          nullable: true
      required:
        - status
...
    CreateLoanProductRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [home, personal, business, education, vehicle, gold]
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
        minTenureMonths:
          type: integer
        maxTenureMonths:
          type: integer
        baseInterestRate:
          type: number
        processingFeePercent:
          type: number
        eligibilityCriteria:
          type: array
          items:
            type: string
        documentsRequired:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required:
        - name
        - type
        - description
        - minAmount
        - maxAmount
        - minTenureMonths
        - maxTenureMonths
        - baseInterestRate
        - processingFeePercent
        - eligibilityCriteria
        - documentsRequired
        - isActive