# lib/api-zod/src/generated/types/createApplicationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Loan Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApplicationRequest {
  loanProductId: number;
  loanAmount: number;
  tenureMonths: number;
  monthlyEmi: number;
  annualIncome?: number;
  employmentType?: string;
}