# lib/api-zod/src/generated/types/createLoanProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Loan Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLoanProductRequestType } from "./createLoanProductRequestType";

export interface CreateLoanProductRequest {
  name: string;
  type: CreateLoanProductRequestType;
  description: string;
  minAmount: number;
  maxAmount: number;
  minTenureMonths: number;
  maxTenureMonths: number;
  baseInterestRate: number;
  processingFeePercent: number;
  eligibilityCriteria: string[];
  documentsRequired: string[];
  isActive: boolean;
}