# lib/api-zod/src/generated/types/emiCalculationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Loan Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface EmiCalculationResult {
  monthlyEmi: number;
  totalAmount: number;
  totalInterest: number;
  processingFee: number;
  loanAmount: number;
  tenureMonths: number;
  interestRate: number;
}