# lib/api-zod/src/generated/types/loanApplication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Loan Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { LoanApplicationAgentStage } from "./loanApplicationAgentStage";
import type { LoanApplicationStatus } from "./loanApplicationStatus";

export interface LoanApplication {
  id: number;
  userId: number;
  loanProductId: number;
  loanProductName: string;
  loanAmount: number;
  tenureMonths: number;
  monthlyEmi: number;
  status: LoanApplicationStatus;
  agentStage?: LoanApplicationAgentStage;
  lockingId?: string | null;
  creditScore?: number | null;
  remarks?: string | null;
  applicantName: string;
  applicantEmail: string;
  applicantPhone: string;
  annualIncome?: number | null;
  employmentType?: string | null;
  createdAt: Date;
  updatedAt: Date;
}