# lib/api-zod/src/generated/types/updateStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Loan Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateStatusRequestAgentStage } from "./updateStatusRequestAgentStage";
import type { UpdateStatusRequestStatus } from "./updateStatusRequestStatus";

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
  agentStage?: UpdateStatusRequestAgentStage;
  remarks?: string | null;
  lockingId?: string | null;
}